# controllers/payment.go
package controllers

import (
	"log"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/ix-pay/ixpay/container"
	"github.com/ix-pay/ixpay/utils"
)

type paymentController struct {
	ctr *container.Container
}

func NewPaymentController(ctr *container.Container) *paymentController {
	return &paymentController{
		ctr: ctr,
	}
}

// CreatePayment
// @Summary 创建支付定单
// @Description 创建支付定单
// @Tags 支付模块
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param userId path string true "定单ID"
// @Success 200 {object} utils.RespData{data=models.ProfileUser} "成功响应"
// @Failure 400 {object} utils.RespData{data=string} "失败消息"
// @Router /payments/create [post]
func (con *paymentController) CreatePayment(c *gin.Context) {
	log.Printf("controllers=%s\n", "创建支付定单")
	// 实现创建支付逻辑
	utils.Success(c, http.StatusOK, "", "创建支付定单")
}

// GetPayment
// @Summary 获取支付定单
// @Description 获取支付定单
// @Tags 支付模块
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param id path string true "定单ID"
// @Success 200 {object} utils.RespData{data=models.ProfileUser} "成功响应"
// @Failure 400 {object} utils.RespData{data=string} "失败消息"
// @Router /payments/{id} [get]
func (con *paymentController) GetPayment(c *gin.Context) {
	log.Printf("controllers=%s\n", "获取支付定单")
	// 实现获取支付详情
	utils.Success(c, http.StatusOK, "", "获取支付定单")
}

// ListPayments
// @Summary 获取支付定单列表
// @Description 获取支付定单列表
// @Tags 支付模块
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param id path string true "定单ID"
// @Success 200 {object} utils.RespData{data=models.ProfileUser} "成功响应"
// @Failure 400 {object} utils.RespData{data=string} "失败消息"
// @Router /payments/list [get]
func (con *paymentController) ListPayments(c *gin.Context) {
	log.Printf("controllers=%s\n", "获取支付定单列表")
	// 实现支付列表查询
	utils.Success(c, http.StatusOK, "", "获取支付定单列表")
}

// UpdatePaymentStatus
// @Summary 更新支付定单状态
// @Description 更新支付定单状态
// @Tags 支付模块
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param id path string true "定单ID"
// @Success 200 {object} utils.RespData{data=models.ProfileUser} "成功响应"
// @Failure 400 {object} utils.RespData{data=string} "失败消息"
// @Router /payments/{id}/status [put]
func (con *paymentController) UpdatePaymentStatus(c *gin.Context) {
	log.Printf("controllers=%s\n", "更新支付定单状态")
	// 实现支付状态更新
	utils.Success(c, http.StatusOK, "", "更新支付定单状态")
}