# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.LoginCredentials:
    properties:
      account:
        minLength: 4
        type: string
      password:
        minLength: 8
        type: string
    required:
    - account
    - password
    type: object
  models.ProfileUser:
    properties:
      account:
        minLength: 4
        type: string
      id:
        type: string
      name:
        minLength: 2
        type: string
    required:
    - account
    - id
    - name
    type: object
  models.RegisterUser:
    properties:
      account:
        minLength: 4
        type: string
      name:
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - account
    - name
    - password
    type: object
  models.TokenUser:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.ProfileUser'
    type: object
  utils.RespData:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
host: localhost:8989
info:
  contact: {}
  description: ixpay项目Swagger集成
  title: API文档
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 登录
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  $ref: '#/definitions/models.TokenUser'
              type: object
        "400":
          description: 失败消息
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  type: string
              type: object
      summary: 登录
      tags:
      - 基础功能
  /auth/profile:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  $ref: '#/definitions/models.ProfileUser'
              type: object
        "400":
          description: 失败消息
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - 基础功能
  /auth/register:
    post:
      consumes:
      - application/json
      description: 注册
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegisterUser'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  $ref: '#/definitions/models.ProfileUser'
              type: object
        "400":
          description: 失败消息
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  type: string
              type: object
      summary: 注册
      tags:
      - 基础功能
  /payments/{id}:
    get:
      consumes:
      - application/json
      description: 获取支付定单
      parameters:
      - description: 定单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  $ref: '#/definitions/models.ProfileUser'
              type: object
        "400":
          description: 失败消息
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: 获取支付定单
      tags:
      - 支付模块
  /payments/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新支付定单状态
      parameters:
      - description: 定单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  $ref: '#/definitions/models.ProfileUser'
              type: object
        "400":
          description: 失败消息
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: 更新支付定单状态
      tags:
      - 支付模块
  /payments/create:
    post:
      consumes:
      - application/json
      description: 创建支付定单
      parameters:
      - description: 定单ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  $ref: '#/definitions/models.ProfileUser'
              type: object
        "400":
          description: 失败消息
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: 创建支付定单
      tags:
      - 支付模块
  /payments/list:
    get:
      consumes:
      - application/json
      description: 获取支付定单列表
      parameters:
      - description: 定单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  $ref: '#/definitions/models.ProfileUser'
              type: object
        "400":
          description: 失败消息
          schema:
            allOf:
            - $ref: '#/definitions/utils.RespData'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: 获取支付定单列表
      tags:
      - 支付模块
securityDefinitions:
  BearerAuth:
    in: header [Bearer ]
    name: Authorization
    type: apiKey
swagger: "2.0"